[not verified] publicize: refresh connections only when post is published
