+++ projects/plugins/jetpack/extensions/plugins/publicize/components/panel/index.js
@@ -19,13 +19,13 @@ import PublicizeConnectionVerify from '../connection-verify';
 import PublicizeForm from '../form';
 import PublicizeTwitterOptions from '../twitter/options';
 import useSelectSocialMediaConnections from '../../hooks/use-social-media-connections';
-import usePostJustSaved from '../../hooks/use-post-just-saved';
+import { usePostJustPublished } from '../../hooks/use-saving-post';
 
 const PublicizePanel = ( { prePublish } ) => {
 	const { refresh, hasEnabledConnections } = useSelectSocialMediaConnections();
 
 	// Refresh connections when the post is just saved.
-	usePostJustSaved(
+	usePostJustPublished(
 		function () {
 			if ( ! hasEnabledConnections ) {
 				return;
